+++ packages/shared/ReactFeatureFlags.js
@@ -24,6 +24,9 @@ export const enablePersistentReconciler = false;
 // Helps identify side effects in begin-phase lifecycle hooks and setState reducers:
 export const debugRenderPhaseSideEffects = false;
 
+// DEV-mode warn about deprecated, async-unsafe lifecycles; relates to RFC #6:
+export const warnAboutDeprecatedLifecycles = false;
+
 // Only used in www builds.
 export function addUserTimingListener() {
   invariant(false, 'Not implemented.');
